+++ demisto_sdk/tests/integration_tests/validate_integration_test.py
@@ -975,7 +975,7 @@ class TestClassifierValidation:
         with ChangeCWD(pack.repo_path):
             runner = CliRunner(mix_stderr=False)
             result = runner.invoke(main, [VALIDATE_CMD, '-i', classifier.path], catch_exceptions=False)
-        assert 'The file type is not supported in validate command' in result.stdout
+        assert 'The file type is not supported in the validate command.' in result.stdout
         assert result.exit_code == 1
 
     def test_valid_old_classifier(self, mocker, repo):
@@ -1228,7 +1228,7 @@ class TestMapperValidation:
         with ChangeCWD(pack.repo_path):
             runner = CliRunner(mix_stderr=False)
             result = runner.invoke(main, [VALIDATE_CMD, '-i', mapper.path], catch_exceptions=False)
-        assert 'The file type is not supported in validate command' in result.stdout
+        assert 'The file type is not supported in the validate command.' in result.stdout
         assert result.exit_code == 1
 
 
@@ -2420,7 +2420,7 @@ class TestImageValidation:
         with ChangeCWD(pack.repo_path):
             runner = CliRunner(mix_stderr=False)
             result = runner.invoke(main, [VALIDATE_CMD, '-i', NOT_VALID_IMAGE_PATH], catch_exceptions=False)
-        assert 'The file type is not supported in validate command' in result.stdout
+        assert 'The file type is not supported in the validate command.' in result.stdout
         assert result.exit_code == 1