+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 * Skipped existence of TestPlaybook check on community and contribution integrations.
 * Fixed an issue where pre-commit didn't run on the demisto_sdk/commands folder.
 * The **init** command will now change the script template name in the code to the given script name.
+* Expanded the validations performed on beta integrations.
+* Improved the error messages in **generate-docs**, if an example was not provided.
+* Added to **validate** command a validation that a content entity or a pack name does not contain the words "partner" and "community".
 
 # 1.4.5
 * Enhanced the **postman-codegen** command to name all generated arguments with lower case.